@@ -12,9 +12,9 @@
 		}
 	],
 	"require": {
-		"php": ">=5.3.0",
-		"predis/predis": "~1.0",
-		"slim/slim": "~2.6"
+		"php": ">=5.4.0",
+		"slim/slim": "~2.6",
+		"predis/predis": "~1.0"
 	},
 	"autoload": {
 		"psr-4": {
